Keep separate package name for simpleapp and composeapp
