+++ noty-android/app/simpleapp/build.gradle
@@ -26,7 +26,7 @@ android {
     buildToolsVersion ProjectConfig.buildToolsVersion
 
     defaultConfig {
-        applicationId "dev.shreyaspatil.noty"
+        applicationId "dev.shreyaspatil.noty.simpleapp"
         minSdkVersion ProjectConfig.minSdkVersion
         targetSdkVersion ProjectConfig.targetSdkVersion
         versionCode ProjectConfig.versionCode
