--- noty-android/app/simpleapp/src/main/java/dev/shreyaspatil/noty/view/base/BaseFragment.kt
@@ -1,57 +1,0 @@
-/*
- * Copyright 2020 Shreyas Patil
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.shreyaspatil.noty.view.base
-
-import android.content.Context
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModel
-import androidx.viewbinding.ViewBinding
-
-abstract class BaseFragment<VB : ViewBinding, VM : ViewModel> : Fragment() {
-
-    private var _binding: VB? = null
-    protected val binding get() = _binding!!
-
-    protected abstract val viewModel: VM
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = getViewBinding(inflater, container)
-        return binding.root
-    }
-
-    protected abstract fun getViewBinding(inflater: LayoutInflater, container: ViewGroup?): VB
-
-    fun toast(message: String) {
-        Toast.makeText(activity, message, Toast.LENGTH_SHORT).show()
-    }
-
-    fun applicationContext(): Context = requireActivity().applicationContext
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
