--- noty-android/app/simpleapp/src/main/java/dev/shreyaspatil/noty/view/home/HomeFragment.kt
@@ -1,50 +1,0 @@
-/*
- * Copyright 2020 Shreyas Patil
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.shreyaspatil.noty.view.home
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import dagger.hilt.android.AndroidEntryPoint
-import dev.shreyaspatil.noty.R
-import dev.shreyaspatil.noty.databinding.HomeFragmentBinding
-import dev.shreyaspatil.noty.view.base.BaseFragment
-import dev.shreyaspatil.noty.view.viewmodel.HomeViewModel
-
-/**
- * Currently nothing is going to performed in HomeFragment.
- */
-@AndroidEntryPoint
-class HomeFragment : BaseFragment<HomeFragmentBinding, HomeViewModel>() {
-
-    override val viewModel: HomeViewModel by viewModels()
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        // Currently doing nothing in Home fragment, so navigate to login screen.
-        findNavController().navigate(R.id.action_homeFragment_to_loginFragment)
-    }
-
-    override fun getViewBinding(
-        inflater: LayoutInflater,
-        container: ViewGroup?
-    ) = HomeFragmentBinding.inflate(inflater, container, false)
-}
