+++ noty-android/app/src/main/AndroidManifest.xml
@@ -14,4 +14,8 @@
   ~ limitations under the License.
   -->
 
-<manifest package="dev.shreyaspatil.noty" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="dev.shreyaspatil.noty">
+
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+</manifest>
